+++ src/fobi/contrib/apps/drf_integration/form_elements/content/content_video/__init__.py
@@ -1,0 +1,11 @@
+__title__ = 'fobi.contrib.apps.drf_integration.form_elements.content.' \
+            'content_video'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = '2014-2017 Artur Barseghyan'
+__license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = ('default_app_config', 'UID',)
+
+default_app_config = 'fobi.contrib.apps.drf_integration.form_elements.' \
+                     'content.content_video.apps.Config'
+
+UID = 'content_video'